Fix the build on Melodic.
When using package version 2, we really only need to use
the <depend> tag, which will mark it as build, exec, and
build_export depend for us.  We also don't need to mark
it separately as test, as that will inherit from the base.
Also make sure to find_package() the packages we need, and
remove the redundant ones in the test.  With all of this,
the tests seem to compile under melodic.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>