Silence cmake warning regarding deprecated FindEigen
PCL searches for a FindEigen.cmake file. The FindEigen.cmake file delivered with PCL actually finds Eigen3, but there is a FindEigen.cmake in cmake_modules that is deprecated and generates a cmake warning. With the old order of the find_packages, there seems to be a mix-up and PCL finds the deprecated file. The new order solves that problem.