Add workaround for deprecated filter functions (#285)
* Add workaround for deprecated filter functions
See also:
https://github.com/PointCloudLibrary/pcl/commit/f26c6fcc24386c29213b62423710c8927cbe1d21
https://github.com/PointCloudLibrary/pcl/commit/630627b2accad395f80bc6518f8dcd94a1f9ad31

* Silence cmake warning regarding deprecated FindEigen

PCL searches for a FindEigen.cmake file. The FindEigen.cmake file delivered with PCL actually finds Eigen3, but there is a FindEigen.cmake in cmake_modules that is deprecated and generates a cmake warning. With the old order of the find_packages, there seems to be a mix-up and PCL finds the deprecated file. The new order solves that problem.